scripts/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0
gpiozero==2.0
python-snap7